pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = ["manh"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
PyMySQL = "^1.0.2"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
